+++ src/controllers/health/index.ts
@@ -4,6 +4,10 @@ import { Logger } from 'pino'
 import * as Ex from 'express'
 import type { Health } from '../../models'
 import { logger } from '../../lib/logger'
+import { startStatusHandlers } from '../../lib/ServiceWatcher/index'
+import { serviceState } from '../../lib/util/statusPoll'
+
+
 @Route('health')
 export class health extends Controller {
   constructor() {
@@ -14,9 +18,30 @@ export class health extends Controller {
   public async get(@Request() req: Ex.Request & { log: Logger; req_id: string }): Promise<Health> {
     logger.debug({ msg: 'new request received', controller: '/health' })
 
-    return Promise.resolve({
-      message: 'success',
-      req_id: req.req_id,
-    })
+    const serviceStatusStrings = {
+      [serviceState.UP]: 'ok',
+      [serviceState.DOWN]: 'down',
+      [serviceState.ERROR]: 'error',
+    }
+
+    const statusHandler = await startStatusHandlers()
+    const status = statusHandler.status
+    const details = statusHandler.detail
+    const code = status === serviceState.UP ? 200 : 503
+    return Promise.resolve(
+      res.status(code).send({
+        version: process.env.npm_package_version ? process.env.npm_package_version : 'unknown',
+        status: serviceStatusStrings[status] || 'error',
+        details: Object.fromEntries(
+          Object.entries(details).map(([depName, { status, detail }]) => [
+            depName,
+            {
+              status: serviceStatusStrings[status] || 'error',
+              detail,
+            },
+          ])
+        ),
+      })
+    )
   }
 }
